requirements.txt
google-genai
Pillow
numpy==1.26.0
pandas==2.1.0
tenacity

